working on person. going to try integrating with a building in a bit

+++ src/mainCity/restaurants/marcusRestaurant/interfaces/Cashier.java
@@ -3,9 +3,9 @@ import mainCity.interfaces.MainCashier;
 import mainCity.restaurants.marcusRestaurant.MarcusTable;
 
 public interface Cashier extends MainCashier {
-	public abstract void msgPayingMyDebt(Customer c, int amount);
+	public abstract void msgPayingMyDebt(Customer c, double amount);
 	
 	public abstract void msgComputeBill(Waiter w, String choice, MarcusTable t);
 	
-	public abstract void msgHereIsPayment(Customer c, int cash, int table);
+	public abstract void msgHereIsPayment(Customer c, double cash, int table);
 }
